app/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlmodel[asyncio]
asyncpg
dnspython
nacl-middleware